+++ src/common/components/modals/DateTimePicker/DateTimePicker.types.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface DateTimePickerProps {
+  selectedDateTime: Date | null;
+  setSelectedDateTime: Dispatch<SetStateAction<Date | null>>;
+}
